requirements.txt
bcrypt==3.2.2
certifi==2021.5.30
cffi==1.15.1
chardet==4.0.0
click==8.0.1
dataclasses==0.8
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT==0.3.2
idna==2.10
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
pkg-resources==0.0.0
protobuf==3.17.3
pycparser==2.21
pycryptodome==3.10.1
PyJWT==1.4.2
requests==2.25.1
six==1.16.0
tinyec==0.3.1
typing-extensions==4.1.1
urllib3==1.26.6
uWSGI==2.0.19.1
Werkzeug==2.0.1
zipp==3.6.0
